Move shopping cart related serializers to payments app
